fix cyclic dependencies for AutoDiscoveryActionEnum
