+++ src/autodiscovery.ts
@@ -17,12 +17,20 @@ limitations under the License.
 
 /** @module auto-discovery */
 
-import { IClientWellKnown, IWellKnownConfig, AutoDiscoveryAction } from "./client";
+import { IClientWellKnown, IWellKnownConfig } from "./client";
 import { logger } from './logger';
 import { URL as NodeURL } from "url";
 
 // Dev note: Auto discovery is part of the spec.
 // See: https://matrix.org/docs/spec/client_server/r0.4.0.html#server-discovery
+
+export enum AutoDiscoveryAction {
+    SUCCESS = "SUCCESS",
+    IGNORE = "IGNORE",
+    PROMPT = "PROMPT",
+    FAIL_PROMPT = "FAIL_PROMPT",
+    FAIL_ERROR = "FAIL_ERROR",
+}
 
 /**
  * Utilities for automatically discovery resources, such as homeservers
